Addressing Jeremy's Feedback.
