+++ packages/front-end/components/SchemaBrowser/PendingInformationSchemaCard.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import LoadingSpinner from "./LoadingSpinner";
+import LoadingSpinner from "../LoadingSpinner";
 
 export default function PendingInformationSchemaCard({
   mutate,
