+++ packages/front-end/components/SchemaBrowser/RetryInformationSchemaCard.tsx
@@ -1,17 +1,20 @@
 import { useEffect, useState } from "react";
+import { InformationSchemaInterface } from "@/../back-end/src/types/Integration";
 import { useAuth } from "@/services/auth";
-import LoadingSpinner from "./LoadingSpinner";
+import LoadingSpinner from "../LoadingSpinner";
 
-export default function BuildInformationSchemaCard({
+export default function RetryInformationSchemaCard({
   datasourceId,
   mutate,
+  informationSchema,
 }: {
   datasourceId: string;
   mutate: () => void;
+  informationSchema: InformationSchemaInterface;
 }) {
-  const [fetching, setFetching] = useState(false);
   const [error, setError] = useState<null | string>(null);
   const { apiCall } = useAuth();
+  const [fetching, setFetching] = useState(false);
   const [retryCount, setRetryCount] = useState(1);
 
   async function onClick() {
@@ -21,11 +24,13 @@ export default function BuildInformationSchemaCard({
         status: number;
         message?: string;
       }>(`/datasource/${datasourceId}/schema`, {
-        method: "POST",
+        method: "PUT",
+        body: JSON.stringify({
+          informationSchemaId: informationSchema.id,
+        }),
       });
       setFetching(true);
     } catch (e) {
-      setFetching(false);
       setError(e.message);
     }
   }
@@ -52,7 +57,7 @@ export default function BuildInformationSchemaCard({
 
   return (
     <div>
-      <div className="alert alert-info">
+      <div className="alert alert-warning d-flex align-items-center">
         <div>
           {fetching ? (
             <span>
@@ -60,19 +65,15 @@ export default function BuildInformationSchemaCard({
               This may take a minute, depending on the size of the datasource.
             </span>
           ) : (
-            <span>
-              Need help building your query? Click the button below to get
-              insight into what tables and columns are available in the
-              datasource.
-            </span>
+            <span>{informationSchema.error.message}</span>
           )}
         </div>
         <button
           disabled={fetching}
-          className="mt-2 btn btn-primary"
+          className="btn btn-link"
           onClick={async () => onClick()}
         >
-          {fetching && <LoadingSpinner />} Generate Information Schema
+          {fetching && <LoadingSpinner />} Retry
         </button>
       </div>
       {error && <div className="alert alert-danger">{error}</div>}
